+++ api/observerrmm/observerrmm/settings.py
@@ -33,16 +33,16 @@ LOST_MODE_EVIDENCE_BASE_PATH = "/opt/observer/lostmode/evidence"
 AUTH_USER_MODEL = "accounts.User"
 
 # latest release
-ORMM_VERSION = "1.4.6"
+ORMM_VERSION = "1.4.7"
 
-WEB_VERSION = "0.2.0"
+WEB_VERSION = "0.2.1"
 
 # bump this version everytime vue code is changed
 # to alert user they need to manually refresh their browser
-APP_VER = "0.1.0"
+APP_VER = "0.1.1"
 
 # https://github.com/braincorp-cl/observer-agent-dist/releases
-LATEST_AGENT_VER = "2.15.14"
+LATEST_AGENT_VER = "2.15.15"
 
 MESH_VER = "1.1.32"